requirements.txt
anyio==3.6.2
bcrypt==4.0.1
click==8.1.3
decorator==5.1.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.92.0
h11==0.14.0
httptools==0.5.0
idna==3.4
loguru==0.6.0
pyasn1==0.4.8
pydantic==1.10.5
pymongo==4.3.3
python-dotenv
python-jose==3.3.0
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
